third_party/WebKit/LayoutTests/http/tests/inspector/network/network-request-type.html

<html>
<head>
<script src="/inspector/inspector-test.js"></script>
<script>
function sendScriptRequest() {
    var script = document.createElement("script");
    script.src = "resources/empty-script.js";
    document.head.appendChild(script);
}

function sendXHRRequest() {
    var xhr = new XMLHttpRequest();
    xhr.open("GET", "resources/empty.html");
    xhr.send();
}

function test() {
    function step1()
    {
        requestName = "empty.html";
        nextStep = InspectorTest.completeTest;
        InspectorTest.evaluateInPage("sendXHRRequest()");
    }

    function onRequest(event)
    {
        var request = event.data;
        if (request.name() !== requestName)
            return;
        requestName = undefined;
        InspectorTest.addResult("");
        InspectorTest.addResult("Request: " + request.name());
        InspectorTest.addResult("Type: " + request.resourceType().name());
        nextStep();
    }

    WebInspector.networkManager.addEventListener(WebInspector.NetworkManager.EventTypes.RequestStarted, onRequest);

    var requestName = "empty-script.js";
    var nextStep = step1;
    InspectorTest.evaluateInPage("sendScriptRequest()");
}
</script>
</head>
<body onload="runTest()">
<p>Tests that XHR request type is detected on send.</p>
</body>
</html>
