third_party/WebKit/LayoutTests/http/tests/inspector/network/network-request-type-expected.txt

Tests that XHR request type is detected on send.


Request: empty-script.js
Type: other

Request: empty.html
Type: xhr
